requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pytest>=7.0.0
pyyaml>=6.0.0 